change u8 to usize for specifying width and height of the Layout2D in Board2D.
Fixing bug that prevents conversion of const generic parameters W and H from u8 to usize to set the sizes of the array used in Layout2D.

+++ src/game_2d/board2d/layout.rs
@@ -1,13 +1,13 @@
-use std::ptr::write_unaligned;
-
-pub struct Layout2D<const W: u8, const H: u8> {
-    layout: [[u16; H as usize]; W as usize],
+pub struct Layout2D<const W: usize, const H: usize>
+{
+    layout: [[u16; H]; W],
 }
 
-impl <const W: u8, const H: u8> Layout2D<W, H> {
+impl <const W: usize, const H: usize> Layout2D<W, H>
+{
     pub fn new() -> Self {
         Self {
-            layout: [[0; H as usize]; W as usize]
+            layout: [[0; H]; W]
         }
     }
